Removed ownjsonSerializer from WindowManager
Removed the unneeded serializer and applied the settings to the global
this._jsonSerializer as when the X and Y values are -1 in the options
it would lead to some settings not being applied. E.g. ContextIsolation
with it now applying the default values when serializing, settings
mentioned before are correctly assigned on Electron's end